+++ event-store-metric-meters/src/main/java/uk/gov/justice/services/eventstore/metrics/meters/gauges/StreamMetricsProvider.java
@@ -1,14 +1,10 @@
 package uk.gov.justice.services.eventstore.metrics.meters.gauges;
 
 import static java.lang.String.format;
-import static java.util.Optional.ofNullable;
-import static java.util.stream.Collectors.toMap;
 
 import uk.gov.justice.services.event.buffer.core.repository.metrics.StreamMetrics;
 import uk.gov.justice.services.event.buffer.core.repository.metrics.StreamMetricsRepository;
 
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import javax.inject.Inject;
@@ -23,17 +19,14 @@ public class StreamMetricsProvider {
     @Inject
     private Logger logger;
 
-    public Optional<StreamMetrics> getMetrics(final String component) {
+    public Optional<StreamMetrics> getMetrics(final String source, final String component) {
 
-        final List<StreamMetrics> streamMetricsList = streamMetricsRepository.getStreamMetrics();
+        final Optional<StreamMetrics> streamMetrics = streamMetricsRepository.getStreamMetrics(source, component);
 
         if (logger.isDebugEnabled()) {
-            logger.debug(format("Retrieved stream metrics for %s: %s", component, streamMetricsList));
+            logger.debug(format("Retrieved stream metrics for %s %s: %s", source, component, streamMetrics));
         }
 
-        final Map<String, StreamMetrics> map = streamMetricsList.stream()
-                .collect(toMap(StreamMetrics::component, streamMetrics -> streamMetrics));
-
-        return ofNullable(map.get(component));
+        return streamMetrics;
     }
 }