+++ event-store-metric-meters/src/test/java/uk/gov/justice/services/eventstore/metrics/meters/gauges/StreamMetricsProviderTest.java
@@ -1,7 +1,5 @@
 package uk.gov.justice.services.eventstore.metrics.meters.gauges;
 
-import static java.util.Arrays.asList;
-import static java.util.Optional.empty;
 import static java.util.Optional.of;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -9,12 +7,9 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.justice.services.core.annotation.Component.EVENT_LISTENER;
-import static uk.gov.justice.services.core.annotation.Component.EVENT_PROCESSOR;
 
 import uk.gov.justice.services.event.buffer.core.repository.metrics.StreamMetrics;
 import uk.gov.justice.services.event.buffer.core.repository.metrics.StreamMetricsRepository;
-import uk.gov.justice.services.eventstore.metrics.meters.gauges.StreamMetricsProvider;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -38,50 +33,36 @@ public class StreamMetricsProviderTest {
     @Test
     public void shouldFindTheCorrectMetricsForComponent() throws Exception {
 
+        final String source = "some-source";
+        final String component = "some-component";
+
         final StreamMetrics eventListenerStreamMetrics = new StreamMetrics(
-                "hearing",
-                EVENT_LISTENER, 234, 329, 23, 87, 78);
-        final StreamMetrics eventProcessorStreamMetrics = new StreamMetrics(
-                "hearing",
-                EVENT_PROCESSOR, 23, 64, 25, 477, 34);
+                source,
+                component, 234, 329, 23, 87, 78);
 
         when(logger.isDebugEnabled()).thenReturn(true);
-        when(streamMetricsRepository.getStreamMetrics()).thenReturn(asList(eventListenerStreamMetrics, eventProcessorStreamMetrics));
+        when(streamMetricsRepository.getStreamMetrics(source, component)).thenReturn(of(eventListenerStreamMetrics));
 
-        assertThat(streamMetricsProvider.getMetrics(eventListenerStreamMetrics.component()), is(of(eventListenerStreamMetrics)));
+        assertThat(streamMetricsProvider.getMetrics(source, component), is(of(eventListenerStreamMetrics)));
 
-        verify(logger).debug("Retrieved stream metrics for EVENT_LISTENER: [StreamMetrics[source=hearing, component=EVENT_LISTENER, streamCount=234, upToDateStreamCount=329, outOfDateStreamCount=23, blockedStreamCount=87, unblockedStreamCount=78], StreamMetrics[source=hearing, component=EVENT_PROCESSOR, streamCount=23, upToDateStreamCount=64, outOfDateStreamCount=25, blockedStreamCount=477, unblockedStreamCount=34]]");
+        verify(logger).debug("Retrieved stream metrics for some-source some-component: Optional[StreamMetrics[source=some-source, component=some-component, streamCount=234, upToDateStreamCount=329, outOfDateStreamCount=23, blockedStreamCount=87, unblockedStreamCount=78]]");
     }
 
     @Test
     public void shouldNotLogIfLogNotOnDebug() throws Exception {
 
+        final String source = "some-source";
+        final String component = "some-component";
+
         final StreamMetrics eventListenerStreamMetrics = new StreamMetrics(
-                "hearing",
-                EVENT_LISTENER, 234, 329, 23, 87, 78);
-        final StreamMetrics eventProcessorStreamMetrics = new StreamMetrics(
-                "hearing",
-                EVENT_PROCESSOR, 23, 64, 25, 477, 34);
+                source,
+                component, 234, 329, 23, 87, 78);
 
         when(logger.isDebugEnabled()).thenReturn(false);
-        when(streamMetricsRepository.getStreamMetrics()).thenReturn(asList(eventListenerStreamMetrics, eventProcessorStreamMetrics));
+        when(streamMetricsRepository.getStreamMetrics(source, component)).thenReturn(of(eventListenerStreamMetrics));
 
-        assertThat(streamMetricsProvider.getMetrics(eventListenerStreamMetrics.component()), is(of(eventListenerStreamMetrics)));
+        assertThat(streamMetricsProvider.getMetrics(source, component), is(of(eventListenerStreamMetrics)));
 
         verify(logger, never()).debug(anyString());
-    }
-
-    @Test
-    public void shouldReturnEmptyIfNoMetricsFoundForComponent() throws Exception {
-        final StreamMetrics eventListenerStreamMetrics = new StreamMetrics(
-                "hearing",
-                EVENT_LISTENER, 234, 329, 23, 87, 78);
-        final StreamMetrics eventProcessorStreamMetrics = new StreamMetrics(
-                "hearing",
-                EVENT_PROCESSOR, 23, 64, 25, 477, 34);
-
-        when(streamMetricsRepository.getStreamMetrics()).thenReturn(asList(eventListenerStreamMetrics, eventProcessorStreamMetrics));
-
-        assertThat(streamMetricsProvider.getMetrics("some-other-component"), is(empty()));
     }
 }