+++ event-store-metric-meters/src/test/java/uk/gov/justice/services/eventstore/metrics/meters/statistic/StreamStatisticTimerBeanTest.java
@@ -1,7 +1,5 @@
 package uk.gov.justice.services.eventstore.metrics.meters.statistic;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;